backend/docker/loader/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
boto3==1.34.0
